% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numberToLetter.R
\name{numberToLetter}
\alias{numberToLetter}
\title{Convert numbers to letters as on a telephone's keypad}
\usage{
numberToLetter(value, decreasing = FALSE)
}
\arguments{
\item{value}{An input value as a character vector with one element (a string)}

\item{decreasing}{Whether to sort the results in alphabetical order or not}
}
\value{
A character vector of letters and dashes based on value
}
\description{
Take a character vector (i.e., a telephone number) and convert it to all
all possible letter combinations as on from a telephone's key pad
}
\examples{
# Convert a string or a vector of numeric characters

numberToLetter("911")
x <- "911"
numberToLetter(x)

# Convert a number directly

numberToLetter(911)

# Convert an alphanumeric string (letters are returned as is and
# non-alphanumeric characters are returned as dashes)

numberToLetter("dial 911!")
}

