% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phonenumber.R
\docType{package}
\name{phonenumber}
\alias{phonenumber}
\alias{phonenumber-package}
\title{phonenumber: Convert letters to numbers and back as on a telephone keypad}
\description{
phonenumber: Convert letters to numbers and back as on a telephone keypad
}
\section{phonenumber functions}{

\itemize{
 \item \code{\link{letterToNumber}}: convert letters to numbers
 \item \code{\link{numberToLetter}}: convert numbers to letters
}
}

