% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/letterToNumber.R
\name{letterToNumber}
\alias{letterToNumber}
\title{Convert letters to numbers as on a telephone's keypad}
\usage{
letterToNumber(value)
}
\arguments{
\item{value}{An input value as a character vector with one element (a string)}
}
\value{
A character vector of numbers and dashes based on value
}
\description{
Take a character vector and convert it to the equivalent number sequence
from a telephone's key pad
}
\examples{
letterToNumber("R functions")
string <- "Phone Number"
letterToNumber(string)
letterToNumber("Jenny's Number is 867-5309")
}

